scripts/replace-certificates/client_requirements.txt
fabric==2.5.0
PyYAML==3.10
